+++ applications/fishing-map/src/data/workspaces.ts
@@ -21,6 +21,9 @@ export enum WorkspaceCategories {
 
 // Different ids for each different environment
 // TODO migrate to use strings and define ids for the following ones
+// IMPORTANT: When updating this list, also update it's corresponding
+// values in:
+//      applications/vessel-history/src/features/dataviews/dataviews.config.ts
 export const DEFAULT_BASEMAP_DATAVIEW_ID = WORKSPACE_ENV === 'development' ? 90 : 173
 export const DEFAULT_VESSEL_DATAVIEW_ID = WORKSPACE_ENV === 'development' ? 92 : 171
 export const DEFAULT_FISHING_DATAVIEW_ID = WORKSPACE_ENV === 'development' ? 91 : 178