+++ applications/vessel-history/src/data/config.ts
@@ -1,48 +1,12 @@
-import { Generators } from '@globalfishingwatch/layer-composer'
-import {
-  BackgroundGeneratorConfig,
-  BasemapGeneratorConfig,
-} from '@globalfishingwatch/layer-composer/dist/generators/types'
-import { ContextLayer } from 'types'
 import { AppState } from 'types/redux.types'
 
-export const MAP_BACKGROUND_COLOR = '#163f89'
-
-export const BACKGROUND_LAYER = [
-  {
-    id: 'background',
-    tileset: 'background',
-    description: 'background',
-    type: Generators.Type.Background,
-    color: MAP_BACKGROUND_COLOR,
-  } as BackgroundGeneratorConfig,
-]
-export const DEFAULT_DATAVIEWS = [
-  {
-    id: 'landmass',
-    tileset: 'landmass',
-    description: 'landmass',
-    type: Generators.Type.Basemap,
-    basemap: Generators.BasemapType.Default,
-  } as BasemapGeneratorConfig,
-]
-export type ContextualLayerTypes = 'landmass'
-
-export const CONTEXT_LAYERS_IDS: { [key in string]: ContextualLayerTypes } = {
-  landmass: 'landmass',
-}
-
-export const CONTEXT_LAYERS: ContextLayer[] = [
-  {
-    id: CONTEXT_LAYERS_IDS.landmass,
-    label: 'Landmass',
-    color: '#6b67e5',
-    description: 'Landmass',
-  },
-]
+export type WorkspaceEnv = 'development' | 'production'
+export const WORKSPACE_ENV =
+  (process.env.REACT_APP_WORKSPACE_ENV as WorkspaceEnv) ||
+  (process.env.NODE_ENV as WorkspaceEnv) ||
+  'production'
 
 export const DEFAULT_WORKSPACE: AppState = {
-  workspaceDataviews: DEFAULT_DATAVIEWS,
   zoom: 3,
   colorMode: 'all',
   minSpeed: 0,