+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -1,9 +1,21 @@
-import React, { Fragment } from 'react'
+import React, { Fragment, useEffect, useState } from 'react'
+import cx from 'classnames'
 import { useTranslation } from 'react-i18next'
-import { Button } from '@globalfishingwatch/ui-components'
+import { useSelector } from 'react-redux'
+import { DateTime } from 'luxon'
+import { Button, IconButton } from '@globalfishingwatch/ui-components'
 import { VesselWithHistory } from 'types'
-import styles from './Info.module.css'
+import { selectCurrentOfflineVessel } from 'features/vessels/offline-vessels.selectors'
+import { useOfflineVesselsAPI } from 'features/vessels/offline-vessels.hook'
+import { OfflineVessel } from 'types/vessel'
+import {
+  selectDataset,
+  selectTmtId,
+  selectVesselId,
+  selectVesselProfileId,
+} from 'routes/routes.selectors'
 import InfoField, { VesselFieldLabel } from './InfoField'
+import styles from './Info.module.css'
 
 interface InfoProps {
   vessel: VesselWithHistory | null
@@ -14,6 +26,42 @@ interface InfoProps {
 const Info: React.FC<InfoProps> = (props): React.ReactElement => {
   const vessel = props.vessel
   const { t } = useTranslation()
+  const [loading, setLoading] = useState(false)
+  const vesselId = useSelector(selectVesselId)
+  const vesselTmtId = useSelector(selectTmtId)
+  const vesselDataset = useSelector(selectDataset)
+  const vesselProfileId = useSelector(selectVesselProfileId)
+  const offlineVessel = useSelector(selectCurrentOfflineVessel)
+  const {
+    dispatchCreateOfflineVessel,
+    dispatchDeleteOfflineVessel,
+    dispatchFetchOfflineVessel,
+  } = useOfflineVesselsAPI()
+
+  useEffect(() => {
+    dispatchFetchOfflineVessel(vesselProfileId)
+  }, [vesselProfileId, dispatchFetchOfflineVessel])
+
+  const onDeleteClick = async (data: OfflineVessel) => {
+    setLoading(true)
+    await dispatchDeleteOfflineVessel(data.profileId)
+    setLoading(false)
+  }
+  const onSaveClick = async (data: VesselWithHistory) => {
+    setLoading(true)
+    await dispatchCreateOfflineVessel({
+      vessel: {
+        ...data,
+        profileId: data.id,
+        id: vesselId,
+        dataset: vesselDataset,
+        vesselMatchId: vesselTmtId,
+        source: '',
+        savedOn: DateTime.utc().toISO(),
+      },
+    })
+    setLoading(false)
+  }
 
   return (
     <Fragment>
@@ -108,12 +156,35 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
             ></InfoField>
           </div>
         )}
-        <br />
-        {vessel && (
-          <Button className={styles.saveButton} type="secondary">
-            {t('vessel.saveForOfflineView', 'SAVE VESSEL FOR OFFLINE VIEW')}
-          </Button>
-        )}
+        <div className={styles.actions}>
+          {vessel && offlineVessel && (
+            <Fragment>
+              <div className={styles.readyForOffline}>
+                {t('vessel.readyForOfflineView', 'READY FOR OFFLINE VIEW')}
+              </div>
+              <IconButton
+                size="default"
+                icon="delete"
+                type="warning"
+                className={cx(styles.defaultIcon, styles.remove)}
+                loading={loading}
+                tooltip={t('vessel.remove', 'Remove')}
+                tooltipPlacement={'top'}
+                onClick={() => onDeleteClick(offlineVessel)}
+              />
+            </Fragment>
+          )}
+          {vessel && !offlineVessel && (
+            <Button
+              className={styles.saveButton}
+              type="secondary"
+              disabled={loading}
+              onClick={() => onSaveClick(vessel)}
+            >
+              {t('vessel.saveForOfflineView', 'SAVE VESSEL FOR OFFLINE VIEW')}
+            </Button>
+          )}
+        </div>
       </div>
     </Fragment>
   )