+++ applications/vessel-history/src/index.tsx
@@ -1,6 +1,7 @@
 import { render } from 'react-dom'
 import React from 'react'
 import { Provider } from 'react-redux'
+import GFWAPI from '@globalfishingwatch/api-client'
 import * as serviceWorker from './serviceWorkerRegistration'
 import store from './store'
 import App from './App'
@@ -19,4 +20,7 @@ render(
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.register()
+serviceWorker.register({
+  // Force performing a login request to cache the response in the SW
+  onSuccess: () => GFWAPI.login({}),
+})