+++ applications/vessel-history/src/serviceWorkerRegistration.ts
@@ -49,6 +49,21 @@ export function register(config?: Config) {
               'worker. To learn more, visit https://cra.link/PWA'
           )
         })
+
+        // TO LISTEN FOR UPDATES ON CACHED RESOURCES
+        // navigator.serviceWorker.addEventListener('message', async (event) => {
+        //   // Optional: ensure the message came from workbox-broadcast-update
+        //   if (event.data.meta === 'workbox-broadcast-update') {
+        //     const { cacheName, updatedUrl } = event.data.payload
+
+        //     // Do something with cacheName and updatedUrl.
+        //     // For example, get the cached content and update
+        //     // the content on the page.
+        //     const cache = await caches.open(cacheName)
+        //     const updatedResponse = await cache.match(updatedUrl)
+        //     const updatedText = await updatedResponse.text()
+        //   }
+        // })
       } else {
         // Is not localhost. Just register service worker
         registerValidSW(swUrl, config)